+++ tests/test_builtin_int.py
@@ -2,7 +2,11 @@ import builtins
 import sys
 import unittest
 
-from builtin_tests.py35 import support
+import builtin_support as support
+from builtin_grammar import (
+    INVALID_UNDERSCORE_LITERALS,
+    VALID_UNDERSCORE_LITERALS,
+)
 
 from fastnumbers import int
 
@@ -214,6 +218,28 @@ class IntTestCases(unittest.TestCase):
         self.assertEqual(int("2br45qc", 35), 4294967297)
         self.assertEqual(int("1z141z5", 36), 4294967297)
 
+    @unittest.skipUnless(
+        sys.version_info >= (3, 6), "Underscores introduced in Python 3.6"
+    )
+    def test_underscores(self):
+        for lit in VALID_UNDERSCORE_LITERALS:
+            if any(ch in lit for ch in ".eEjJ"):
+                continue
+            self.assertEqual(int(lit, 0), eval(lit))
+            self.assertEqual(int(lit, 0), int(lit.replace("_", ""), 0))
+        for lit in INVALID_UNDERSCORE_LITERALS:
+            if any(ch in lit for ch in ".eEjJ"):
+                continue
+            self.assertRaises(ValueError, int, lit, 0)
+        # Additional test cases with bases != 0, only for the constructor:
+        self.assertEqual(int("1_00", 3), 9)
+        self.assertEqual(int("0_100"), 100)  # not valid as a literal!
+        self.assertEqual(int(b"1_00"), 100)  # byte underscore
+        self.assertRaises(ValueError, int, "_100")
+        self.assertRaises(ValueError, int, "+_100")
+        self.assertRaises(ValueError, int, "1__00")
+        self.assertRaises(ValueError, int, "100_")
+
     @support.cpython_only
     def test_small_ints(self):
         # Bug #3236: Return small longs from PyLong_FromString
@@ -227,9 +253,15 @@ class IntTestCases(unittest.TestCase):
 
     def test_keyword_args(self):
         # Test invoking int() using keyword arguments.
-        self.assertEqual(int(x=1.2), 1)
         self.assertEqual(int("100", base=2), 4)
-        self.assertEqual(int(x="100", base=2), 4)
+        if sys.version_info >= (3, 7):
+            with self.assertRaisesRegex(TypeError, "keyword argument"):
+                int(x=1.2)
+            with self.assertRaisesRegex(TypeError, "keyword argument"):
+                int(x="100", base=2)
+        else:
+            self.assertEqual(int(x=1.2), 1)
+            self.assertEqual(int(x="100", base=2), 4)
         self.assertRaises(TypeError, int, base=10)
         self.assertRaises(TypeError, int, base=0)
 
@@ -357,6 +389,22 @@ class IntTestCases(unittest.TestCase):
 
             self.assertEqual(int(IntOverridesTrunc()), 42)
 
+    @unittest.skipUnless(sys.version_info >= (3, 8), "Test introduced in Python 3.8")
+    def test_int_subclass_with_index(self):
+        class MyIndex(builtins.int):
+            def __index__(self):
+                return 42
+
+        class BadIndex(builtins.int):
+            def __index__(self):
+                return 42.0
+
+        my_int = MyIndex(7)
+        self.assertEqual(my_int, 7)
+        self.assertEqual(int(my_int), 7)
+
+        self.assertEqual(int(BadIndex()), 0)
+
     def test_int_subclass_with_int(self):
         class MyInt(builtins.int):
             def __int__(self):
@@ -370,9 +418,22 @@ class IntTestCases(unittest.TestCase):
         self.assertEqual(my_int, 7)
         self.assertEqual(int(my_int), 42)
 
-        self.assertRaises(TypeError, int, BadInt())
+        if sys.version_info >= (3, 8):
+            my_int = BadInt(7)
+            self.assertEqual(my_int, 7)
+            self.assertRaises(TypeError, int, my_int)
+        else:
+            self.assertRaises(TypeError, int, BadInt())
 
     def test_int_returns_int_subclass(self):
+        class BadIndex:
+            def __index__(self):
+                return True
+
+        class BadIndex2(builtins.int):
+            def __index__(self):
+                return True
+
         class BadInt:
             def __int__(self):
                 return True
@@ -381,15 +442,31 @@ class IntTestCases(unittest.TestCase):
             def __int__(self):
                 return True
 
+        if sys.version_info >= (3, 8):
+            bad_int = BadIndex()
+            with self.assertWarns(DeprecationWarning):
+                n = int(bad_int)
+            self.assertEqual(n, 1)
+            self.assertIs(type(n), builtins.int)
+
+            bad_int = BadIndex2()
+            n = int(bad_int)
+            self.assertEqual(n, 0)
+            self.assertIs(type(n), builtins.int)
+
         bad_int = BadInt()
         with self.assertWarns(DeprecationWarning):
             n = int(bad_int)
         self.assertEqual(n, 1)
+        if sys.version_info >= (3, 6):
+            self.assertIs(type(n), builtins.int)
 
         bad_int = BadInt2()
         with self.assertWarns(DeprecationWarning):
             n = int(bad_int)
         self.assertEqual(n, 1)
+        if sys.version_info >= (3, 6):
+            self.assertIs(type(n), builtins.int)
 
     def test_error_message(self):
         def check(s, base=None):
@@ -424,6 +501,16 @@ class IntTestCases(unittest.TestCase):
         # lone surrogate in Unicode string
         check("123\ud800")
         check("123\ud800", 10)
+
+    @unittest.skipUnless(sys.version_info >= (3, 7), "Test introduced in Python 3.7")
+    def test_issue31619(self):
+        self.assertEqual(
+            int("1_0_1_0_1_0_1_0_1_0_1_0_1_0_1_0_1_0_1_0_1_0_1_0_1_0_1_0_1_0_1", 2),
+            0b1010101010101010101010101010101,
+        )
+        self.assertEqual(int("1_2_3_4_5_6_7_0_1_2_3", 8), 0o12345670123)
+        self.assertEqual(int("1_2_3_4_5_6_7_8_9", 16), 0x123456789)
+        self.assertEqual(int("1_2_3_4_5_6_7", 32), 1144132807)
 
 
 if __name__ == "__main__":
